requirements.txt
# =========================
# CORE DEEP LEARNING
# =========================
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.2.0

# =========================
# NUMERICAL & IMAGE PROCESSING
# =========================
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.10.0

# =========================
# DATA AUGMENTATION
# =========================
albumentations>=1.3.1

# =========================
# MACHINE LEARNING / METRICS
# =========================
scikit-learn>=1.3.0

# =========================
# VISUALIZATION
# =========================
matplotlib>=3.7.0
seaborn>=0.12.0

# =========================
# LOGGING & EXPERIMENT TRACKING
# =========================
tensorboard>=2.13.0
wandb>=0.16.0
tqdm>=4.66.0

# =========================
# CONFIG & FILE HANDLING
# =========================
PyYAML>=6.0
pathlib2>=2.3.7

# =========================
# MODEL EXPORT / DEPLOYMENT
# =========================
onnx>=1.14.0
onnxruntime>=1.16.0

# =========================
# GAN E
